More work on menus
Added in the ability to quit to main menu, and added a menu button to
the main game. Tends to throw errors if you try to play the game
directly from the scene now, start from main menu instead. Also going to
menu then returning has the side effect of resetting the level. Perhaps
use a flat pause instead while in game?